Moved predicate factory into its own file. Removed observation because swiftData import gets observation by default
